Porting findNeighbours on GPU
