+++ Detectors/ITSMFT/ITS/tracking/GPU/cuda/CMakeLists.txt
@@ -13,13 +13,14 @@
 if(CUDA_ENABLED)
 find_package(CUDAToolkit)
 message(STATUS "Building ITS CUDA tracker")
-
+# add_compile_options(-O0 -g -lineinfo -fPIC)
 o2_add_library(ITStrackingCUDA
                SOURCES ClusterLinesGPU.cu
                        Context.cu
                        Stream.cu
                        TrackerTraitsGPU.cxx
                        TimeFrameGPU.cu
+                       TimeFrameChunk.cu
                        TracerGPU.cu
                        TrackingKernels.cu
                        VertexingKernels.cu
@@ -31,7 +32,7 @@ o2_add_library(ITStrackingCUDA
                                      O2::SimulationDataFormat
                                      O2::ReconstructionDataFormats
                                      O2::GPUCommon
-                                     CUDA::nvToolsExt # TODO: change to CUDA::nvtx3 when CMake bump >= 3.25
+                                     CUDA::nvToolsExt
                PRIVATE_LINK_LIBRARIES O2::GPUTrackingCUDAExternalProvider
                TARGETVARNAME targetName)
