+++ Detectors/ITSMFT/ITS/tracking/include/ITStracking/TimeFrame.h
@@ -288,6 +288,7 @@ class TimeFrame
   std::vector<Road<5>> mRoads;
   std::vector<std::vector<TrackITSExt>> mTracks;
   std::vector<std::vector<int>> mCellsNeighbours;
+  std::vector<std::vector<int>> mCellsLookupTable;
 
   const o2::base::PropagatorImpl<float>* mPropagatorDevice = nullptr; // Needed only for GPU
  protected:
@@ -315,7 +316,6 @@ class TimeFrame
   std::vector<std::array<float, 2>> mPValphaX; /// PV x and alpha for track propagation
   std::vector<std::vector<MCCompLabel>> mTrackletLabels;
   std::vector<std::vector<MCCompLabel>> mCellLabels;
-  std::vector<std::vector<int>> mCellsLookupTable;
   std::vector<std::vector<int>> mCellsNeighboursLUT;
   std::vector<std::vector<MCCompLabel>> mTracksLabel;
   std::vector<int> mBogusClusters; /// keep track of clusters with wild coordinates
